// src/state/types.ts
export interface DualFarm extends DualFarmConfig {
  tokenAmount?: BigNumber
  totalInQuoteToken?: BigNumber
  quoteTokenAmount?: BigNumber
  lpTotalInQuoteToken?: BigNumber
  tokenPriceVsQuote?: BigNumber
  stakeTokenPrice?: number
  rewardToken0Price?: number
  rewardToken1Price?: number
  poolWeight?: BigNumber
  multiplier?: string
  apr?: number
  apy?: string
  lpApr?: string
  totalStaked?: string
  userData?: {
    allowance: BigNumber
    tokenBalance: BigNumber
    stakedBalance: BigNumber
    miniChefEarnings: BigNumber
    rewarderEarnings: BigNumber
  }
}
// ...
export interface FarmLpAprsType {
  chainId: number
  lpAprs: {
    pid: number
    lpApr: number
  }[]
}
// ...
export interface TokenPrices {
  symbol: string
  address: Address
  price: number
  decimals: Decimals
}

export interface LpTokenPrices {
  symbol: string
  pid: number
  address: Address
  price: number
  decimals: number
}

// Slices states

export interface ToastsState {
  data: Toast[]
}

export interface FarmsState {
  data: Farm[]
}

export interface PoolsState {
  data: Pool[]
}

export interface NfaState {
  isInitialized: boolean
  isLoading: boolean
  data: Nfa[]
}

export interface DualFarmsState {
  data: DualFarm[]
}